rename: 아코디언 파츠 컴포넌트명 & 타입명 단순화(#45)
